Nueva estructuración de archivos.
