+++ pyplexos/reader/pcp_model/xml_model.py
@@ -1,0 +1,1 @@
+# TODO. Crear la estructura con pydantic-xml el modelo de entrada a plexos.
