+++ pyplexos/reader/zip_model/bin_model.py
